refactor: Refactor Spartan polynomial file organization
- Refactor the structure of Spartan SNARK related modules, mainly moving and renaming `polynomial.rs` to `polys/mlpoly.rs` and introducing three new submodules under `polys`: `eq_poly`, `mlpoly`, and `unipoly`.
- Implement a new module `polys` that houses refactored and newly introduced polynomial types used in the Spartan SNARK, leading to import refactors across numerous files.
- Establish a new file `unipoly.rs` responsible for housing the main structs `UniPoly` and `CompressedUniPoly`
