+++ src/spartan/mod.rs
@@ -7,14 +7,14 @@
 //! In polynomial.rs we also provide foundational types and functions for manipulating multilinear polynomials.
 pub mod direct;
 pub(crate) mod math;
-pub mod polynomial;
+pub mod polys;
 pub mod ppsnark;
 pub mod snark;
 mod sumcheck;
 
 use crate::{traits::Group, Commitment};
 use ff::Field;
-use polynomial::SparsePolynomial;
+use polys::mlpoly::SparsePolynomial;
 
 fn powers<G: Group>(s: &G::Scalar, n: usize) -> Vec<G::Scalar> {
   assert!(n >= 1);
