+++ src/spartan/polys/mod.rs
@@ -1,0 +1,4 @@
+//! This module contains the definitions of polynomial types used in the Spartan SNARK.
+pub mod eq_poly;
+pub mod mlpoly;
+pub(crate) mod unipoly;
