+++ src/spartan/sumcheck.rs
@@ -1,9 +1,12 @@
 #![allow(clippy::too_many_arguments)]
 #![allow(clippy::type_complexity)]
-use super::polynomial::MultilinearPolynomial;
 use crate::errors::NovaError;
-use crate::traits::{Group, TranscriptEngineTrait, TranscriptReprTrait};
-use ff::{Field, PrimeField};
+use crate::spartan::polys::{
+  mlpoly::MultilinearPolynomial,
+  unipoly::{CompressedUniPoly, UniPoly},
+};
+use crate::traits::{Group, TranscriptEngineTrait};
+use ff::Field;
 use rayon::prelude::*;
 use serde::{Deserialize, Serialize};
 
@@ -297,111 +300,5 @@ impl<G: Group> SumcheckProof<G> {
       r,
       vec![poly_A[0], poly_B[0], poly_C[0], poly_D[0]],
     ))
-  }
-}
-
-// ax^2 + bx + c stored as vec![a,b,c]
-// ax^3 + bx^2 + cx + d stored as vec![a,b,c,d]
-#[derive(Debug)]
-pub struct UniPoly<Scalar: PrimeField> {
-  coeffs: Vec<Scalar>,
-}
-
-// ax^2 + bx + c stored as vec![a,c]
-// ax^3 + bx^2 + cx + d stored as vec![a,c,d]
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct CompressedUniPoly<Scalar: PrimeField> {
-  coeffs_except_linear_term: Vec<Scalar>,
-}
-
-impl<Scalar: PrimeField> UniPoly<Scalar> {
-  pub fn from_evals(evals: &[Scalar]) -> Self {
-    // we only support degree-2 or degree-3 univariate polynomials
-    assert!(evals.len() == 3 || evals.len() == 4);
-    let two_inv = Scalar::from(2).invert().unwrap();
-    let coeffs = if evals.len() == 3 {
-      // ax^2 + bx + c
-      let c = evals[0];
-      let a = two_inv * (evals[2] - evals[1] - evals[1] + c);
-      let b = evals[1] - c - a;
-      vec![c, b, a]
-    } else {
-      // ax^3 + bx^2 + cx + d
-      let six_inv = Scalar::from(6).invert().unwrap();
-
-      let d = evals[0];
-      let a = six_inv
-        * (evals[3] - evals[2] - evals[2] - evals[2] + evals[1] + evals[1] + evals[1] - evals[0]);
-      let b = two_inv
-        * (evals[0] + evals[0] - evals[1] - evals[1] - evals[1] - evals[1] - evals[1]
-          + evals[2]
-          + evals[2]
-          + evals[2]
-          + evals[2]
-          - evals[3]);
-      let c = evals[1] - d - a - b;
-      vec![d, c, b, a]
-    };
-
-    UniPoly { coeffs }
-  }
-
-  pub fn degree(&self) -> usize {
-    self.coeffs.len() - 1
-  }
-
-  pub fn eval_at_zero(&self) -> Scalar {
-    self.coeffs[0]
-  }
-
-  pub fn eval_at_one(&self) -> Scalar {
-    (0..self.coeffs.len())
-      .into_par_iter()
-      .map(|i| self.coeffs[i])
-      .sum()
-  }
-
-  pub fn evaluate(&self, r: &Scalar) -> Scalar {
-    let mut eval = self.coeffs[0];
-    let mut power = *r;
-    for coeff in self.coeffs.iter().skip(1) {
-      eval += power * coeff;
-      power *= r;
-    }
-    eval
-  }
-
-  pub fn compress(&self) -> CompressedUniPoly<Scalar> {
-    let coeffs_except_linear_term = [&self.coeffs[0..1], &self.coeffs[2..]].concat();
-    assert_eq!(coeffs_except_linear_term.len() + 1, self.coeffs.len());
-    CompressedUniPoly {
-      coeffs_except_linear_term,
-    }
-  }
-}
-
-impl<Scalar: PrimeField> CompressedUniPoly<Scalar> {
-  // we require eval(0) + eval(1) = hint, so we can solve for the linear term as:
-  // linear_term = hint - 2 * constant_term - deg2 term - deg3 term
-  pub fn decompress(&self, hint: &Scalar) -> UniPoly<Scalar> {
-    let mut linear_term =
-      *hint - self.coeffs_except_linear_term[0] - self.coeffs_except_linear_term[0];
-    for i in 1..self.coeffs_except_linear_term.len() {
-      linear_term -= self.coeffs_except_linear_term[i];
-    }
-
-    let mut coeffs: Vec<Scalar> = Vec::new();
-    coeffs.push(self.coeffs_except_linear_term[0]);
-    coeffs.push(linear_term);
-    coeffs.extend(&self.coeffs_except_linear_term[1..]);
-    assert_eq!(self.coeffs_except_linear_term.len() + 1, coeffs.len());
-    UniPoly { coeffs }
-  }
-}
-
-impl<G: Group> TranscriptReprTrait<G> for UniPoly<G::Scalar> {
-  fn to_transcript_bytes(&self) -> Vec<u8> {
-    let coeffs = self.compress().coeffs_except_linear_term;
-    coeffs.as_slice().to_transcript_bytes()
   }
 }
